Remove webdriver extsions
